Update AcornSatTransferFunctionAnalysis - but I won't even use it...
