+++ ClimateExplorer.Data.AcornSatTransferFunctionAnalysis/AcornSatFileParser.cs
