+++ ClimateExplorer.Data.AcornSatTransferFunctionAnalysis/HqNewFileParser.cs
